microkernel2/requirements.txt
termcolor
jinja2
pyyaml
